requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0
Werkzeug==3.0.3
